Fix misleading typo
My understanding is that the edited sentence only makes sense with the proposed change. A node does not know about its public IP, private to public NAT happens on the InternetGateways. Also, nodes can only have a single public IP, a single primary private IP, and many secondary private IPs. So it doesn't make sense to talk about a 'primary public IP', but it is meaningful to say 'primary private IP'.

Unfortunately, the attached image is also misleading, as it shows a private to public SNAT done by the CNI plugin, whereas it is a SNAT between a secondary private to the primary private address. The primary private to public SNAT happens on the gateway, this could be also shown.