Clarify custom networks for multi-AZ node groups
There was some confusion when following the instructions for setting up
custom network configuration for multiple Availability Zones. The previous
documentation included instructions to use the `--node-labels` bootstrap
argument when creating a new worker node group, and specified the AZ
name as a node label. This doesn't work, however, when the user wants to
have a worker node group in multiple AZs. Instead, the availability zone of
the node is already added in the `failure-domain.beta.kubernetes.io/zone`
label.

We clarify these instructions by removing the `--node-labels` part of the bootstrap
args in the step which creates a worker node group and add some notes to step
5a and 5c that indicate recommended practice is to set the ENIConfig's `name`
field to that of the Availability Zone and then simply issue the call to:

```
kubectl set env daemonset aws-node -n kube-system ENI_CONFIG_LABEL_DEF=failure-domain.beta.kubernetes.io/zone
```

which will be appropriate even for worker node groups in multiple AZs.