Update private-clusters.md
AWS Loadbalancer Controller works in the private cluster / AirGapped cluster but with some limitations. The limitations are not specified in the AWS EKS Private cluster documentations yet but are discussed in the forums.
https://github.com/kubernetes-sigs/aws-load-balancer-controller/issues/1855

We could either update all the limitations in this current considerations section or we could just mention its supported and create a "Considerations for Private Cluster" in the AWS Loadbalancer controller installation page.