Update security-groups-for-pods.md
Putting brackets around <role> makes it sound like a key that needs to be substituted, while in reality the policy yaml needs to contain the actual key named "role", without any brackets.  Since I used the sample policy as a reference, it took me a while to figure out the typo in the key, as the failure mode is very subtle.