Add some notes on verifying the IAM role
Customers were occasionally getting confused when creating an IAM role with the `eksctl` command. The most common issue was that a role with that name already existed and so the `eksctl` command did not do anything. This change adds a verification step after the role is created. It also recommends using the cluster name in the role name.