Removes "AWS Management Console" section from launching self-managed nodes
The [AWS Management Console section](https://docs.aws.amazon.com/eks/latest/userguide/launch-workers.html) references a launch template with a 1.17 nodegroup, which is long deprecated. If customer feedback/comments hasn't motivated us to change this already, we can safely assume that customers don't depend on this and can remove the burden of maintaining it altogether.