Updating formatting and removes `No`
Update formatting as the words are being merged into one so they need a space. Also removes the `No` from one of the FAQ answers because later on in the answer we say that there are no plans. So it's best to just remove the `No` because it gives the impression that there it will never be removed, whereas the `No plans to remove them` doesn't remove that option completely, it just says there are no plans currently (which could change)