Clarify the signerName in the example as being the only valid signerName
I had used this example, but it was not clear to me that the `beta.eks.amazonaws.com/app-serving` value for `signerName` was required in order for EKS to sign my certificates, and I had used my own custom signer name instead. This PR adds some suggestions for clarifying the required value of `signerName` to have EKS sign any certificates.